fix(gates): split the claim check out of the census into a check_ gate
`tests/scripts/gate_population.test.ts` red on CI: "manifest lists
report_skill_activation, population does not". `_lib/gate_population.ts`
classifies a gate by prefix so that "is this a gate" is answered structurally,
and `report_*` is deliberately outside it — registering the census in
`src/config/gate-coverage.yml` forked that population.

Widening the prefix set to admit `report_` was the wrong repair: it would add
eleven reporting scripts to the population, and `gate-hardening:unhardened-scan-scope`
sits at zero with no baseline.

So the verdict moves to where a verdict belongs. `check_skill_activation_claim`
compares the census record against the sentence `docs/CLAIMS.md` publishes,
carries the `scanned:` floor, the `--self-test` and the `ci_invocation` pin; the
census keeps its exit-0 contract and its `--emit` empty-store refusal. The
manifest row, the release-validation step, the locality command, the parity
declaration and the claim's `exec:` pointer all follow.
