+++ docs/CLAIMS.md
@@ -242,9 +242,9 @@ metacharacters and repo escape, including the right-hand side of `--flag=value`.
 - last_verified: 2026-08-26
 
 ### claim: skill-activation-census-zero
-- claim: Skill self-selection is measured, not assumed, and over this package's own transcript store it is ZERO — not near zero. `report_skill_activation` over 30 sessions and 11,338 assistant turns records 0 Skill invocations and 0 of 299 distinct skills (measured 2026-09-06). Every figure in this sentence, and the date, is read out of `agents/evidence/metrics/skill-activation-census.json`, which the census writes and refuses to write from an empty store; `report_skill_activation --check` fails when the record and this sentence disagree, so the published number cannot outlive its measurement silently. The 299 split three ways and the zero means a different thing in each: 12 declare a machine-matchable trigger key in frontmatter, 100 carry an `evals/triggers.json` corpus, 2 do both, and the remaining 189 are reachable only by a human naming them. For the 12 the zero is a defect — a matchable declaration exists and nothing acted on it. For the 189 the zero is the design, because no automatic selection is claimed for them and invocation there is a reader opening a file. For the 100 the zero licenses nothing in either direction: `evals/triggers.json` is a TEST fixture read by `check_routing_coverage` / `lint_skill_trigger_corpus` / `check_trigger_evals`, no host reads it at routing time, and the two surfaces share the word "trigger" and are unrelated. A 14-file eval-corpus wave taking coverage 100 → 114 was authored, measured against, and then reverted on the same branch for an unrelated governance reason; the census read 0 in both states. <!-- count: dated -->
+- claim: Skill self-selection is measured, not assumed, and over this package's own transcript store it is ZERO — not near zero. `report_skill_activation` over 30 sessions and 11,338 assistant turns records 0 Skill invocations and 0 of 299 distinct skills (measured 2026-09-06). Every figure in this sentence, and the date, is read out of `agents/evidence/metrics/skill-activation-census.json`, which the census writes and refuses to write from an empty store; `check_skill_activation_claim` fails when the record and this sentence disagree, so the published number cannot outlive its measurement silently. The 299 split three ways and the zero means a different thing in each: 12 declare a machine-matchable trigger key in frontmatter, 100 carry an `evals/triggers.json` corpus, 2 do both, and the remaining 189 are reachable only by a human naming them. For the 12 the zero is a defect — a matchable declaration exists and nothing acted on it. For the 189 the zero is the design, because no automatic selection is claimed for them and invocation there is a reader opening a file. For the 100 the zero licenses nothing in either direction: `evals/triggers.json` is a TEST fixture read by `check_routing_coverage` / `lint_skill_trigger_corpus` / `check_trigger_evals`, no host reads it at routing time, and the two surfaces share the word "trigger" and are unrelated. A 14-file eval-corpus wave taking coverage 100 → 114 was authored, measured against, and then reverted on the same branch for an unrelated governance reason; the census read 0 in both states. <!-- count: dated -->
 - kind: quant
-- evidence: exec:report_skill_activation --check -> 0
+- evidence: exec:check_skill_activation_claim -> 0
 - non_inference: This is a reading of ONE transcript store on one machine, so it bounds this package's own maintainer sessions and licenses nothing about consumer installs, other hosts, or any period outside the 30 sessions scanned. The store is rolling rather than append-only, so a later re-derivation can legitimately report FEWER assistant turns than an earlier one; a rise or fall between two readings is not a trend. Zero invocations is NOT evidence that the skills are useless, unreachable, or wrongly written — the census names its own leading unmeasured cause, that the host's injected catalogue may carry bare names rather than descriptions past roughly the first forty entries, and that catalogue is not persisted in the transcript so nothing here tests it. The three population counts partition the 299 by what each skill DECLARES, not by what any host does with it: they do not establish that the 189 are unreachable, that the 12 are reachable, or that the split is the right one. The number is also not a rate: with a zero numerator there is no denominator question to answer. It says nothing about whether `lean_projection.mode: delivery` would move activation; that needs the flip, which is owner-reserved. And it must not be read as a verdict on the corpus wave, which raised testability and was never a delivery mechanism.
 - status: backed
 - last_verified: 2026-09-06
